non multipart fileupload

+++ backend/rewind.php
@@ -15,22 +15,25 @@ $admin = $query['USR_ID'];
 
 if($type == 0) {
 	if($admin != $user){
-        $dir = "var/www/html/" . $fac . "/" . $user . "_in/";
+        $dir = "/var/www/html/" . $fac . "/" . $user . "_in/";
 	} else {
-		$dir = "var/www/html/" . $fac . "/" . $user . "/";
+		$dir = "/var/www/html/" . $fac . "/" . $user . "/";
 	}    
 } else {
 	if($admin != $user){
-        $dir = "var/www/html/" . $fac . "/" . $admin . "_alt/";
+        $dir = "/var/www/html/" . $fac . "/" . $admin . "_alt/";
 	} else {
-		$dir = "var/www/html/" . $fac . "/" . $user . "_alt/";
+		$dir = "/var/www/html/" . $fac . "/" . $user . "_alt/";
 	}
 }
 
 $extension = pathinfo($fname, PATHINFO_EXTENSION);
 $var = basename($fname, "." . $extension);
 
-unlink($dir . $var . "_[" . $stsk . "]_" . $user . "." . $extension);
+if( is_file($dir . $var . "_[" . $stsk . "]_" . $user . "." . $extension)){
+	 unlink($dir . $var . "_[" . $stsk . "]_" . $user . "." . $extension);
+}
+
 
 mysqli_close($datos);
